bubble up debezium config error messages as exceptions (#33658)
Looks like the logic of handling the result of the debezium engine was slightly wrong. Wee were assining the local `error` field, and then checking itf it was not null, and overriding it in that case. The truth is that the engine can fail without filling in the `error` field, but still fill in the `message` field. So the new logic uses the `message` field if `error` is null

fixes #31579