cdk-java move the generationId handling to its own class (#43329)
This pull request introduces a new interface, `JdbcGenerationHandler`, to handle the retrieval of the `_airbyte_generation_id` for any row in a table. The `getGenerationIdInTable` method has been moved from `SqlOperations` to this new interface. The necessary changes have been made across various classes to integrate this new interface properly.